% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAddins-package.R
\docType{package}
\name{createAddins-package}
\alias{createAddins}
\alias{createAddins-package}
\title{createAddins: Create commonly used Addins in RStudio}
\description{
Create commonly used Addins in RStudio, including '%<>%'.
}
\author{
\strong{Maintainer}: Zhen Lu \email{luzh29@mail2.sysu.edu.cn} (\href{https://orcid.org/0000-0002-3481-6310}{ORCID})

}
\keyword{internal}
