% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertPipeOperator.R
\name{insertPipeOperator}
\alias{insertPipeOperator}
\title{Insert pipe operator of \code{\%<>\%}}
\usage{
insertPipeOperator(arg = NULL)
}
\arguments{
\item{arg}{NULL. No need to provide an argument.}
}
\description{
This is a function named 'insertPipeOperator'
which inserts \code{\%<>\%} at the cursor position.
}
